SQUASH ME: resolve merge conflicts
