+++ qa/src/test/java/io/camunda/migrator/qa/runtime/jobtype/ConfigureAnotherJobTypeValidationTest.java
@@ -6,30 +6,29 @@
  * except in compliance with the Camunda License 1.0.
  */
 
-package io.camunda.migrator.qa;
+package io.camunda.migrator.qa.runtime.jobtype;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
 import io.camunda.migrator.RuntimeMigrator;
-import io.camunda.migrator.qa.util.RuntimeMigrationAbstractTest;
+import io.camunda.migrator.qa.runtime.RuntimeMigrationAbstractTest;
 import io.github.netmikey.logunit.api.LogCapturer;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.RegisterExtension;
 import org.springframework.test.context.TestPropertySource;
 
 @TestPropertySource(properties = {
-    "camunda.migrator.job-type=custom-activation-type",
-    "camunda.migrator.validation-job-type==if legacyId != null then \"migrator\" else \"noop\""
+    "camunda.migrator.job-type=my-job-type",
 })
-public class SeparateJobTypesValidationTest extends RuntimeMigrationAbstractTest {
+public class ConfigureAnotherJobTypeValidationTest extends RuntimeMigrationAbstractTest {
 
   @RegisterExtension
   protected LogCapturer logs = LogCapturer.create().captureForType(RuntimeMigrator.class);
 
   @Test
-  public void shouldUseValidationJobTypeInValidationMessage() {
+  public void shouldUseAnotherJobTypeInValidationMessage() {
     // given
-    deployProcessInC7AndC8("noMigratorListener.bpmn");
+    deployer.deployProcessInC7AndC8("noMigratorListener.bpmn");
 
     String id = runtimeService.startProcessInstanceByKey("noMigratorListener").getId();
 
@@ -45,16 +44,16 @@ public class SeparateJobTypesValidationTest extends RuntimeMigrationAbstractTest
     var events = logs.getEvents();
     assertThat(events.stream().filter(event -> event.getMessage()
         .matches(String.format(".*Skipping process instance with legacyId \\[%s\\]: "
-            + "No execution listener of type '=if legacyId != null then \"migrator\" else \"noop\"' found on "
+            + "No execution listener of type 'my-job-type' found on "
             + "start event \\[Event_1px2j50\\] in C8 process with id \\[(\\d+)\\]\\. "
-            + "At least one '=if legacyId != null then \"migrator\" else \"noop\"' listener is required\\.", id))))
+            + "At least one 'my-job-type' listener is required\\.", id))))
         .hasSize(1);
   }
 
   @Test
-  public void shouldUseValidationJobTypeInListenerNotFoundMessage() {
-    // given
-    deployProcessInC7AndC8("migratorListenerCustomType.bpmn");
+  public void shouldUseAnotherJobTypeInListenerNotFoundMessage() {
+    // given: BPM model uses 'foo' job type in the listener which doesn't match
+    deployer.deployProcessInC7AndC8("migratorListenerCustomType.bpmn");
 
     String id = runtimeService.startProcessInstanceByKey("migratorListenerCustomType").getId();
 
@@ -70,27 +69,9 @@ public class SeparateJobTypesValidationTest extends RuntimeMigrationAbstractTest
     var events = logs.getEvents();
     assertThat(events.stream().filter(event -> event.getMessage()
         .matches(String.format(".*Skipping process instance with legacyId \\[%s\\]: "
-            + "No execution listener of type '=if legacyId != null then \"migrator\" else \"noop\"' found on "
+            + "No execution listener of type 'my-job-type' found on "
             + "start event \\[Event_1px2j50\\] in C8 process with id \\[(\\d+)\\]\\. "
-            + "At least one '=if legacyId != null then \"migrator\" else \"noop\"' listener is required\\.", id))))
+            + "At least one 'my-job-type' listener is required\\.", id))))
         .hasSize(1);
   }
-
-  @Test
-  public void shouldUseCustomValidationJobTypeInListenerSucceed() {
-    // given
-    deployProcessInC7AndC8("migratorListenerFeel.bpmn");
-
-    String id = runtimeService.startProcessInstanceByKey("migratorListenerFeel").getId();
-
-    // assume
-    assertThat(runtimeService.createProcessInstanceQuery().processInstanceId(id).singleResult()).isNotNull();
-
-    // when
-    runtimeMigrator.start();
-
-    // then
-    assertThatProcessInstanceCountIsEqualTo(1);
-  }
-
 }
