+++ qa/src/test/java/io/camunda/migrator/qa/runtime/jobtype/ExternalTrafficTest.java
@@ -6,12 +6,12 @@
  * except in compliance with the Camunda License 1.0.
  */
 
-package io.camunda.migrator.qa;
+package io.camunda.migrator.qa.runtime.jobtype;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
 import io.camunda.migrator.RuntimeMigrator;
-import io.camunda.migrator.qa.util.RuntimeMigrationAbstractTest;
+import io.camunda.migrator.qa.runtime.RuntimeMigrationAbstractTest;
 import io.github.netmikey.logunit.api.LogCapturer;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.RegisterExtension;
@@ -24,7 +24,7 @@ public class ExternalTrafficTest extends RuntimeMigrationAbstractTest {
   @Test
   public void shouldHandleExternallyStartedMigratorJobsGracefully() {
     // given
-    deployProcessInC7AndC8("simpleProcess.bpmn");
+    deployer.deployProcessInC7AndC8("simpleProcess.bpmn");
 
     String id = runtimeService.startProcessInstanceByKey("simpleProcess").getId();
