+++ qa/src/test/java/io/camunda/migrator/qa/runtime/jobtype/SeparateJobTypesValidationTest.java
@@ -6,29 +6,30 @@
  * except in compliance with the Camunda License 1.0.
  */
 
-package io.camunda.migrator.qa;
+package io.camunda.migrator.qa.runtime.jobtype;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
 import io.camunda.migrator.RuntimeMigrator;
-import io.camunda.migrator.qa.util.RuntimeMigrationAbstractTest;
+import io.camunda.migrator.qa.runtime.RuntimeMigrationAbstractTest;
 import io.github.netmikey.logunit.api.LogCapturer;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.RegisterExtension;
 import org.springframework.test.context.TestPropertySource;
 
 @TestPropertySource(properties = {
+    "camunda.migrator.job-type=custom-activation-type",
     "camunda.migrator.validation-job-type==if legacyId != null then \"migrator\" else \"noop\""
 })
-public class ValidationJobTypeOnlyTest extends RuntimeMigrationAbstractTest {
+public class SeparateJobTypesValidationTest extends RuntimeMigrationAbstractTest {
 
   @RegisterExtension
   protected LogCapturer logs = LogCapturer.create().captureForType(RuntimeMigrator.class);
 
   @Test
-  public void shouldUseCustomValidationJobTypeWithDefaultActivationType() {
+  public void shouldUseValidationJobTypeInValidationMessage() {
     // given
-    deployProcessInC7AndC8("noMigratorListener.bpmn");
+    deployer.deployProcessInC7AndC8("noMigratorListener.bpmn");
 
     String id = runtimeService.startProcessInstanceByKey("noMigratorListener").getId();
 
@@ -51,9 +52,9 @@ public class ValidationJobTypeOnlyTest extends RuntimeMigrationAbstractTest {
   }
 
   @Test
-  public void shouldUseCustomValidationJobTypeInListenerNotFoundMessage() {
+  public void shouldUseValidationJobTypeInListenerNotFoundMessage() {
     // given
-    deployProcessInC7AndC8("migratorListenerCustomType.bpmn");
+    deployer.deployProcessInC7AndC8("migratorListenerCustomType.bpmn");
 
     String id = runtimeService.startProcessInstanceByKey("migratorListenerCustomType").getId();
 
@@ -78,7 +79,7 @@ public class ValidationJobTypeOnlyTest extends RuntimeMigrationAbstractTest {
   @Test
   public void shouldUseCustomValidationJobTypeInListenerSucceed() {
     // given
-    deployProcessInC7AndC8("migratorListenerFeel.bpmn");
+    deployer.deployProcessInC7AndC8("migratorListenerFeel.bpmn");
 
     String id = runtimeService.startProcessInstanceByKey("migratorListenerFeel").getId();
